settings reorganization
